+++ 2_8/src/main/java/com/aleks/_8/DAL/Filters/UpdateContactFilter.java
@@ -22,9 +22,9 @@ public class UpdateContactFilter {
         return ContactPhoneNumber;
     }
 
-    private Long ContactId;
+    private int ContactId;
 
-    public Long getContactId() {
+    public int getContactId() {
         return ContactId;
     }
 }
